+++ lib/webview_gtk_embed_method_channel.dart
@@ -11,7 +11,8 @@ class MethodChannelWebviewGtkEmbed extends WebviewGtkEmbedPlatform {
 
   @override
   Future<String?> getPlatformVersion() async {
-    final version = await methodChannel.invokeMethod<String>('getPlatformVersion');
+    final version =
+        await methodChannel.invokeMethod<String>('getPlatformVersion');
     return version;
   }
 }